refactor: extract Renderer struct from ApplicationState
Move GPU rendering fields (surface, device, queue, surface_config,
pipeline, uniform_buffer, bind_group) from ApplicationState into a
dedicated Renderer struct in src/renderer.rs. This groups wgpu
rendering concerns into a cohesive unit while keeping fields pub
for minimal API churn.

Closes #98
